[GPT] Generated Suggestions ## Goal Add comments to only classes.
#### Path: /engineer
